3_script/python/GUI/qt/vap_open_server_SDK/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
paho-mqtt==1.5.1
requests==2.25.1
urllib3==1.26.4
